Update migration configuration and enhance schema management
- Changed MySQL and PostgreSQL connection details in the .env file for updated database access.
- Introduced a new migration progress tracking file, `migration_progress.json`, to store migration statistics.
- Enhanced the `ensure_schema` function to support a force option for schema recreation.
- Improved error handling and logging during schema application and data migration processes.
- Added support for unique constraints and foreign key handling in the migration logic.
- Updated PostgreSQL schema definitions to use appropriate data types and constraints.

+++ migrate/schema/postgres_schema.sql
@@ -16,7 +16,7 @@ CREATE UNIQUE INDEX ix_event_types_event_type ON event_types(event_type);
 -- Table: event_data
 CREATE TABLE event_data (
     data_id BIGSERIAL PRIMARY KEY,
-    hash INTEGER,
+    hash BIGINT,
     shared_data TEXT
 );
 
@@ -49,7 +49,7 @@ CREATE INDEX ix_events_data_id ON events(data_id);
 -- Table: state_attributes
 CREATE TABLE state_attributes (
     attributes_id BIGSERIAL PRIMARY KEY,
-    hash INTEGER,
+    hash BIGINT,
     shared_attrs TEXT
 );
